fix: scheduling error transaction propagation fix
