+++ SchedulingService/src/main/java/com/example/empsched/scheduling/service/ScheduleSolverService.java
@@ -23,30 +23,15 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class ScheduleSolverService {
 
-    private final ShiftRepository shiftRepository;
-    private final EmployeeRepository employeeRepository;
     private final SolverManager<WorkSchedule, UUID> solverManager;
-    private final ScheduleRepository scheduleRepository;
     private final ScheduleService scheduleService;
-
-    private final EmployeeAvailabilityRepository availabilityRepository; // Inject this
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW, readOnly = true)
-    public WorkSchedule loadProblem(final UUID scheduleId) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow();
-        List<Employee> employees = employeeRepository.findAllByOrganisationId(schedule.getOrganisation().getId());
-        List<Shift> shifts = shiftRepository.findAllByScheduleId(scheduleId);
-        List<EmployeeAvailability> availabilities = availabilityRepository
-                .findAllByDateBetween(schedule.getStartDate(), schedule.getEndDate());
-        return new WorkSchedule(scheduleId, employees, shifts, availabilities);
-    }
 
     public void solveSchedule(final UUID scheduleId) {
         scheduleService.setScheduleStatus(scheduleId, ScheduleStatus.SOLVING);
         solverManager.solveBuilder()
                 .withProblemId(scheduleId)
                 // 1. INPUT: Load the problem
-                .withProblemFinder(this::loadProblem)
+                .withProblemFinder(scheduleService::loadProblem)
                 // 2. OUTPUT: Save only the FINAL result
                 .withFinalBestSolutionConsumer(scheduleService::saveSolution)
                 // 3. ERROR HANDLING:
